// apps/admin/src/lib/api/generated/models/adminListCommunitiesStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen Platform Admin API
 * OpenAPI spec version: 0.1.0
 */

export type AdminListCommunitiesStatus = typeof AdminListCommunitiesStatus[keyof typeof AdminListCommunitiesStatus] | null;


export const AdminListCommunitiesStatus = {
  draft: 'draft',
  active: 'active',
  paused: 'paused',
} as const;